Extract mocks
